feat: code clean up
