+++ README.md
@@ -1,46 +1,24 @@
-# Getting Started with Create React App
+# Jeffrey Siu's personal github page
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+## Purpose
 
-## Available Scripts
+This repo contains the source code for my personal github page.
 
-In the project directory, you can run:
+## Built With
 
-### `yarn start`
+- [React](https://reactjs.org/)
+- [create-react-app](https://create-react-app.dev/)
+- [react router](https://reactrouter.com/)
+- [MUI](https://mui.com/)
 
-Runs the app in the development mode.\
-Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
+## Experience
 
-The page will reload if you make edits.\
-You will also see any lint errors in the console.
+I started this project to learn more about React. It was a fun and fulfilling experience as I did not have a ton of front-end development experience.
 
-### `yarn test`
+I initially chose to create the application with [Next.js](https://nextjs.org/), as it's a React framework with a bunch of features, such as image optimization and file-system routing, built in. 
 
-Launches the test runner in the interactive watch mode.\
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
+However, I encountered some issues after I deployed the code to the `gh-pages` branch: Routing did not work, and Next.js' image loader would not work with `next export`.
 
-### `yarn build`
+I tried to fix the problems with these posts: ["Deploy with Travis"](https://dev.to/jameswallis/deploying-a-next-js-app-to-github-pages-24pn) and ["Deploy with GitHub Actions"](https://www.linkedin.com/pulse/deploy-nextjs-app-github-pages-federico-antu%C3%B1a/), but I was not able to fix the issues.
 
-Builds the app for production to the `build` folder.\
-It correctly bundles React in production mode and optimizes the build for the best performance.
-
-The build is minified and the filenames include the hashes.\
-Your app is ready to be deployed!
-
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
-
-### `yarn eject`
-
-**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
-
-If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
-
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
-
-You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
-
-## Learn More
-
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
+In the end, I rewrote the application with plain React and React Router. Fortunately, I did not have to do a ton of refactoring since the code was already split up as components.
