+++ src/components/bio/bio.tsx
@@ -1,0 +1,10 @@
+import Content from "../common/content";
+import Profile from "./profile";
+
+function GetContent() {
+  return <Profile />;
+}
+
+export default function Bio() {
+  return <Content ChipLabel="Info" Content={GetContent()} />;
+}
